Add support for wl_display types for native display
In the nested wayland composition scenario, the compositor needs to forward the
DRM authentication to its server. Only the DRM master can authenticate the
client.

We store the wl_display and pass it to the wl_kms server in the middle, so that
the wl_kms server in the middle can forward the authentication request to its
server.

This fix requires the latest wayland-kms.

Change-Id: Iee7d87f1097270cb328252bd53230bbb0bf62159
Signed-off-by: Tony SIM <chinyeow.sim.xt@renesas.com>